create abstraction for waiting for ui to settle
